Split schema into separate files in a module
Closes #1
