+++ src/schema/types.rs
@@ -1,8 +1,4 @@
 use serde::{Serialize, Deserialize};
-use std::sync::Mutex;
-use juniper::FieldResult;
-
-use crate::database;
 
 #[derive(juniper::GraphQLObject, Serialize, Deserialize, Debug, Clone)]
 pub struct Cocktail {
@@ -92,85 +88,3 @@ impl From<NewRating> for Rating {
         }
     }
 }
-
-#[derive(Debug)]
-pub struct Context {
-    pub db: Mutex<database::Database>,
-}
-
-impl juniper::Context for Context {}
-
-pub struct Query;
-
-#[juniper::object(
-    Context = Context,
-)]
-impl Query {
-    fn apiVersion() -> &str {
-        "0.1"
-    }
-
-    fn cocktail(ctx: &Context, id: i32) -> FieldResult<Cocktail> {
-        let db = ctx.db.lock()?;
-        let cocktail = db.get_cocktail(id)?;
-
-        Ok(cocktail)
-    }
-
-    fn cocktails(context: &Context) -> FieldResult<Vec<Cocktail>> {
-        let db = context.db.lock()?;
-        let mut cocktails = db.get_all_cocktails()?;
-
-        Ok(cocktails)
-    }
-
-    fn users(context: &Context) -> FieldResult<Vec<String>> {
-        let db = context.db.lock()?;
-        let users = db.get_all_users()?;
-
-        Ok(users)
-    }
-}
-
-pub struct Mutation;
-
-#[juniper::object(
-    Context = Context,
-)]
-impl Mutation {
-    fn createCocktail(context: &Context, new_cocktail: NewCocktail) -> FieldResult<Cocktail> {
-        let db = context.db.lock()?;
-        let cocktail = db.add_cocktail(&new_cocktail)?;
-
-        Ok(cocktail)
-    }
-
-    fn editCocktail(context: &Context, id: i32, new_cocktail: NewCocktail) -> FieldResult<Cocktail> {
-        let db = context.db.lock()?;
-        let cocktail = db.overwrite_cocktail(id, &new_cocktail)?;
-
-        Ok(cocktail)
-    }
-
-    fn deleteCocktail(context: &Context, id: i32) -> FieldResult<i32> {
-        let db = context.db.lock()?;
-        db.delete_cocktail(id)?;
-
-        Ok(id)
-    }
-
-    fn rateCocktail(context: &Context, id: i32, rating: NewRating) -> FieldResult<Vec<Rating>> {
-        let db = context.db.lock()?;
-        db.rate_cocktail(id, rating.into())?;
-
-        let ratings = db.get_cocktail(id)?.ratings;
-        Ok(ratings)
-    }
-}
-
-pub type Schema = juniper::RootNode<'static, Query, Mutation>;
-
-pub fn create_schema() -> Schema {
-    Schema::new(Query, Mutation)
-}
-
